feat: перенести в модуль server все классы для работы с бронированиями
