+++ shareIt-server/src/main/java/ru/practicum/shareit/server/booking/dto/BookingShortDto.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.booking.dto;
+package ru.practicum.shareit.server.booking.dto;
 
 import lombok.Data;
