+++ shareIt-server/src/main/java/ru/practicum/shareit/server/booking/dto/PostBookingRequest.java
@@ -1,0 +1,10 @@
+package ru.practicum.shareit.server.booking.dto;
+
+import lombok.Data;
+
+@Data
+public class PostBookingRequest {
+    private Integer itemId;
+    private String start;
+    private String end;
+}
