+++ shareIt-server/src/main/java/ru/practicum/shareit/server/booking/model/Booking.java
@@ -1,9 +1,10 @@
-package ru.practicum.shareit.booking.model;
+package ru.practicum.shareit.server.booking.model;
 
 import jakarta.persistence.*;
 import lombok.*;
-import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.user.model.User;
+
+import ru.practicum.shareit.server.item.model.Item;
+import ru.practicum.shareit.server.user.model.User;
 
 import java.time.LocalDateTime;
